Move some css code to the respective folders
