+++ gatsby-node.ts
@@ -17,7 +17,7 @@ const path = require('path');
 
 // Define the template for blog and blog post
 const blogPost = path.resolve('./src/templates/blog-post/index.tsx');
-const blog = path.resolve('./src/templates/blog.tsx');
+const blog = path.resolve('./src/templates/blog/index.tsx');
 const comparisonTemplate = path.resolve(
     './src/templates/product-comparison.tsx'
 );
