+++ src/components/Bio/styles.module.less
@@ -1,0 +1,11 @@
+.avatarWrapper {
+  align-items: center;
+  display: flex;
+  white-space: nowrap;
+  font-size: 1rem;
+}
+
+.styledLink {
+  align-items: center;
+  display: flex;
+}
